Update top bar buttons based on page
