+++ src/app/layout.tsx
@@ -12,11 +12,13 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body>
+      <body
+        style={{ display: "flex", flexDirection: "column", minHeight: "100vh" }}
+      >
         <ThemeProvider theme={theme}>
           <CssBaseline />
           <TopBar />
-          {children}
+          <main style={{ flex: 1, overflowY: "auto" }}>{children}</main>
           <Footer />
         </ThemeProvider>
       </body>
