+++ src/components/ToskaLogo.tsx
@@ -1,14 +1,16 @@
-"use client";
+"use server";
 
 import { Box, Typography } from "@mui/material";
 import Image from "next/image";
 
-export const ToskaLogo = () => {
+export const ToskaLogo = async () => {
   return (
     <Box alignItems="center" display="flex" flexDirection="column">
       <Image
         alt="Toska logo"
         height={200}
+        loading="eager"
+        priority
         src="./assets/toska-logo.svg"
         title="Toska logo, design by Jukka Päivinen"
         width={300}
